requirements.txt
Django==4.2.7
docker==6.1.3
celery==5.3.4
redis==5.0.1
psutil==5.9.6
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-bootstrap5==23.3
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
djangorestframework==3.14.0
django-allauth==0.57.0
django-guardian==2.4.0
django-extensions==3.2.3
